config: Add multiple SSID support.
Add knowledge of 'wifi-network' blocks in /etc/config/wireless file to
netifd. This separates the wifi network information from the wifi
interface interface making it possible to have multiple wifi networks
tied to a single wifi interface (e.g. when running in client mode having
a list of networks for wpa-supplicant from which to choose).

The each of the 'wifi-network' blocks will be translated into a
'network={}' block in the wpa-supplicant.conf file that is generated by
the 'package/network/services/hostapd/files/netifd.sh' script.

Example block:

  config wifi-network
          option device 'radio0'
          option iface 'wlan0'
          option ssid 'example'
          option encryption 'none'

The 'device' field is not currently used.
The 'iface' field is used to tie this network block to the wifi interface.

The JSON blob that gets passed into the mac80211.sh script by netifd
will now include a 'wnetworks' array element where there is an element
in the array for each 'wifi-network' block in the /etc/config/wireless
file.

Signed-off-by: Theodore A. Roth <theodore_roth@trimble.com>